requirements.txt
kivy
kociemba
pillow
